pyproject.toml
[tool.poetry]
name = "osgameclones"
version = "1.0.0"
description = "Open Source Game Clones"
authors = ["Alexander Solovyov <alexander@solovyov.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
natsort = "8.4.0"
pykwalify = "1.8.0"
pykwalify-webform = "0.2.1"
Unidecode = "1.3.2"
htmlmin = "0.1.12"
jinja2 = "3.1.4"
python-slugify = "^8.0.1"

[tool.poetry.dev-dependencies]
yamllint = "^1.27.1"
beautifulsoup4 = "^4.12.2"
tenacity = "^8.2.2"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
mistletoe = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
